Fix scene graph FBO/shader effect sources
Turns out the scene graph OpenGL ES code path requires
GL_PES_packed_depth_stencil to function properly. Filter
the reported extensions and return it if present.

It's uncertain wether filtering like this does any
good or if we should just hardcode support for it
or fail with a descriptive error message if it's not
present.

A quick web search indicates that Chrome emulates the
feature if the extension is not present (Chrome issue
15793012)