Revert "Fix fallbacks for adapted common script"
This reverts 1dd9a0af4f577ccb5578cea562a98686c8e290e6. It was
a band-aid for a change in the unicode itemizing algorithm
which caused the script of a script item to become unreliable.
This change has since been reverted, so the band-aid is no longer
needed, and it also causes problems for WebKit on Windows when
it ends up preferring Arial Unicode MS as the font for Uchen
script, even though the font does not support this script.

The autotest from the reverted commit is kept in place and still
passes.

[ChangeLog][Text] Fixed regression when rendering Uchen text in
WebKit on Windows.

Change-Id: I488c84703bb55a050d90092c6bf9e5c70a9e31c2
Task-number: QTBUG-41372
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>