Remove mkspecs and detection of Visual Studio .NET 2003
This compiler is no longer supported, as the mkspec was moved to
unsupported/ on commit 55c3799bd32d6fd8ec4b5fa26ebe3e5fdbbc91b3, but the
MSVC support in qmake and in configure depend on an exact string
match. So remove the remaining bits.

No changelog because the actual removal happened in an earlier Qt release.

Change-Id: I538345f4184a6af2ea7449052c161afe4eac625c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>