WIP: Add QPlatformWindow::window* accessors.
The platform plugins should use this to get geometry
and sizes in device pixels. (The corresponding QWindow
accessors return values in device-independent pixels).

This should be easier to use from the platform plugin
implementors point of view: there is no longer a need
to pick and use one of the qHighDpi* functions. We
can also hide implementation details like the "Constrained"
functions and normalization.

Change-Id: I1e42b56ecce99f13b1a3f86a0a8cd05afff06d42