Android: Enable -fno-builtin-memmove in arm builds
On some devices, a miscompilation of libc.so has caused it to
return the wrong value. Instead of returning the dest pointer,
it returns dest + n. When compiling with optimizations turned
on, gcc may use this return value for subsequent accesses to
dest after the memmove() call, causing memory corruption.

This caused problems e.g. in QVector::prepend() which would
overwrite the whole vector with the new value.

Setting -fno-builtin-memmove disables the optimization and
works around this bug with very little risk or impact.

More information in:
     http://code.google.com/p/android/issues/detail?id=81692

[ChangeLog][Android] Fixed device-specific crash on Samsung
Galaxy Tab 3 Lite 7" and some other devices.

Task-number: QTBUG-34984
Change-Id: I0c1347149eb5fe1c298758fe7de81aca4137f652
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>